Moved the entity type configuration classes into more appropriate locations.
